refactor: 기념일 페이지 기능에 따라서 재배치
