+++ src/app/group/anniversary/setting/page.tsx
@@ -1,16 +1,15 @@
 'use client';
 
 import React from 'react';
-import SettingForm from '@/components/anniversary/SettingForm';
 import Header from '@/components/common/Header';
+import SettingForm from '@/components/group/anniversary/setting/SettingForm';
 
-export default function Page() {
+export default function AnniversarySettingPage() {
   return (
     <>
       <Header isShowButton isShowProfile={false}>
         기념일 설정
       </Header>
-      기념일
       <SettingForm />
     </>
   );
