+++ src/components/group/anniversary/setting/SettingForm.module.scss
@@ -11,7 +11,7 @@
     width: 100%;
     border: 0.05rem solid $gray;
     border-radius: 0.6rem;
-    padding: 1.5rem 28.6rem 1.5rem 2.1rem;
+    padding: 1.5rem 0 1.5rem 2.1rem;
     margin-top: 1rem;
     &:focus {
       border-color: $primary-10;
